#292 - Integrate local read preference Commit 3:  * Configure maintenance agents with local read preference  * Rename of scheduledtask to generic
