+++ docs/configmgmt.md
@@ -53,6 +53,6 @@ The [OVS commandline](https://openvstorage.gitbooks.io/administration/content/Ad
 ```
 
 ### Changing the schedule for the ALBA backend verifictaion process
-In case you want to change the schedule for the ALBA backend verifictaion process which checks the state of each object in the backend, add `"alba.scheduled.verify_namespaces": {"minute": "0", "hour": "0", "month_of_year": "*/X"}` to the `/ovs/framework/scheduling/celery` JSON object. In this schedule X is the amount of months between each run.
+In case you want to change the schedule for the ALBA backend verifictaion process which checks the state of each object in the backend, add `"alba.verify_namespaces": {"minute": "0", "hour": "0", "month_of_year": "*/X"}` to the `/ovs/framework/scheduling/celery` JSON object. In this schedule X is the amount of months between each run.
 
 In case the configuration cannot be parsed at all (e.g. invalid JSON), the code will fallback to the hardcoded schedule. If the crontab arguments are invalid (e.g. they contain an unsupported key) the task will be disabled.
